requirements.txt
asgiref==3.8.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.3
djangorestframework==3.15.0
idna==3.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
